move objectlist which lists s3 from prune to common
