+++ internal/prune/graphprune.go
@@ -2,6 +2,7 @@ package prune
 
 import (
 	"fmt"
+	"github.com/gleanerio/nabu/internal/common"
 	"github.com/gleanerio/nabu/internal/graph"
 	"github.com/gleanerio/nabu/internal/objects"
 	"github.com/gleanerio/nabu/pkg/config"
@@ -27,7 +28,7 @@ func Snip(v1 *viper.Viper, mc *minio.Client) error {
 	for p := range pa {
 
 		// do the object assembly
-		oa, err := ObjectList(v1, mc, pa[p])
+		oa, err := common.ObjectList(v1, mc, pa[p])
 		if err != nil {
 			log.Println(err)
 			return err
